+++ MiniEngine/Core/PipelineState.h
@@ -68,7 +68,7 @@ public:
     void SetPrimitiveTopologyType( D3D12_PRIMITIVE_TOPOLOGY_TYPE TopologyType );
     void SetRenderTargetFormat( DXGI_FORMAT RTVFormat, DXGI_FORMAT DSVFormat, UINT MsaaCount = 1, UINT MsaaQuality = 0 );
     void SetRenderTargetFormats( UINT NumRTVs, const DXGI_FORMAT* RTVFormats, DXGI_FORMAT DSVFormat, UINT MsaaCount = 1, UINT MsaaQuality = 0 );
-    void SetInputLayout( UINT NumElements, const D3D12_INPUT_ELEMENT_DESC* pInputElementDescs );
+    void SetInputLayout( size_t NumElements, const D3D12_INPUT_ELEMENT_DESC* pInputElementDescs );
     void SetPrimitiveRestart( D3D12_INDEX_BUFFER_STRIP_CUT_VALUE IBProps );
 
     // These const_casts shouldn't be necessary, but we need to fix the API to accept "const void* pShaderBytecode"
